+++ tests/unit/ProctorioServiceTest.php
@@ -25,6 +25,7 @@ namespace oat\Proctorio\tests\unit;
 use oat\Proctorio\ProctorioConfig;
 use oat\Proctorio\ProctorioAccessProvider;
 use oat\Proctorio\ProctorioService;
+use oat\Proctorio\Response\ProctorioResponse;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 
@@ -42,16 +43,16 @@ class ProctorioServiceTest extends TestCase
     private const SECRET = 'secret';
 
     private const CONFIG_EXAMPLE = [
-        ProctorioConfig::LAUNCH_URL  => self::LAUNCH_URL_CUSTOM_VALUE,
-        ProctorioConfig::USER_ID  => self::USER_ID_CUSTOM_VALUE,
-        ProctorioConfig::OAUTH_CONSUMER_KEY  => self::OAUTH_CONSUMER_KEY_CUSTOM_VALUE,
-        ProctorioConfig::EXAM_START  => self::LAUNCH_URL_CUSTOM_VALUE,
-        ProctorioConfig::EXAM_TAKE  => self::EXAM_TAKE_CUSTOM_VALUE,
+        ProctorioConfig::LAUNCH_URL => self::LAUNCH_URL_CUSTOM_VALUE,
+        ProctorioConfig::USER_ID => self::USER_ID_CUSTOM_VALUE,
+        ProctorioConfig::OAUTH_CONSUMER_KEY => self::OAUTH_CONSUMER_KEY_CUSTOM_VALUE,
+        ProctorioConfig::EXAM_START => self::LAUNCH_URL_CUSTOM_VALUE,
+        ProctorioConfig::EXAM_TAKE => self::EXAM_TAKE_CUSTOM_VALUE,
         ProctorioConfig::EXAM_END => self::EXAM_END_CUSTOM_VALUE,
-        ProctorioConfig::EXAM_SETTINGS  => 'recordaudio,recordvideo',
-        ProctorioConfig::OAUTH_SIGNATURE_METHOD  => 'HMAC-SHA1',
-        ProctorioConfig::OAUTH_VERSION  => '1.0',
-        ProctorioConfig::OAUTH_TIMESTAMP  => '1586522824',
+        ProctorioConfig::EXAM_SETTINGS => 'recordaudio,recordvideo',
+        ProctorioConfig::OAUTH_SIGNATURE_METHOD => 'HMAC-SHA1',
+        ProctorioConfig::OAUTH_VERSION => '1.0',
+        ProctorioConfig::OAUTH_TIMESTAMP => '1586522824',
         ProctorioConfig::OAUTH_NONCE => 'nonce',
     ];
 
@@ -61,9 +62,9 @@ class ProctorioServiceTest extends TestCase
         ProctorioConfig::EXAM_START => self::LAUNCH_URL_CUSTOM_VALUE,
         ProctorioConfig::EXAM_END => self::EXAM_END_CUSTOM_VALUE,
         ProctorioConfig::EXAM_TAKE => self::EXAM_TAKE_CUSTOM_VALUE,
-        ProctorioConfig::OAUTH_TIMESTAMP  => '1586522824',
+        ProctorioConfig::OAUTH_TIMESTAMP => '1586522824',
         ProctorioConfig::OAUTH_NONCE => 'nonce',
-        ProctorioConfig::EXAM_SETTINGS  => self::EXAM_SETTINGS_EXAMPLE,
+        ProctorioConfig::EXAM_SETTINGS => self::EXAM_SETTINGS_EXAMPLE,
     ];
 
     /** @var ProctorioAccessProvider|MockObject */
@@ -82,11 +83,20 @@ class ProctorioServiceTest extends TestCase
 
     public function testCallRemoteProctoring(): void
     {
-        $this->proctorioProviderMock->expects($this->once())
+        $expectedResponse = new ProctorioResponse('url1', 'url2');
+
+        $this->proctorioProviderMock
+            ->expects($this->once())
             ->method('retrieve')
             ->with(self::CONFIG_EXAMPLE, self::SECRET)
-            ->willReturn('string');
+            ->willReturn($expectedResponse);
 
-        $this->subject->callRemoteProctoring(self::PARAMS_EXAMPLE, self::OAUTH_CONSUMER_KEY_CUSTOM_VALUE, self::SECRET);
+        $this->assertSame($expectedResponse,
+            $this->subject->callRemoteProctoring(
+                self::PARAMS_EXAMPLE,
+                self::OAUTH_CONSUMER_KEY_CUSTOM_VALUE,
+                self::SECRET
+            )
+        );
     }
 }